# specs/001-podcast-review-site/contracts/api.yaml
openapi: 3.1.0
info:
  title: Podcast 聽後回顧網站 API
  version: 1.0.0
  description: |
    Podcast 聽後回顧網站的 REST API 規格。
    本 API 主要透過 Next.js API Routes 實作，部分操作直接使用 Supabase Client。
servers:
  - url: /api
    description: Next.js API Routes
...
paths:
...
  /search:
    get:
      tags: [Search]
      summary: 全文搜尋
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: 搜尋關鍵字
        - name: show
          in: query
          schema:
            type: string
            format: uuid
          description: 篩選特定節目系列
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
          description: 篩選標籤
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: 起始日期
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: 結束日期
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /affiliate/redirect/{affiliateId}:
    get:
      tags: [Affiliate]
      summary: 聯盟行銷重導向
      description: 記錄點擊並重導向至目標 URL
      parameters:
        - name: affiliateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: episode
          in: query
          schema:
            type: string
            format: uuid
          description: 來源單集 ID
      responses:
        '302':
          description: 重導向至目標 URL
...
  /admin/comments:
    get:
      tags: [Admin]
      summary: 取得待審留言
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, hidden, spam]
            default: pending
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCommentListResponse'
  /admin/comments/{id}:
    patch:
      tags: [Admin]
      summary: 更新留言狀態
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [approved, hidden, spam]
              required:
                - status
      responses:
        '200':
          description: 更新成功
        '401':
          description: 未登入
        '403':
          description: 權限不足
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT Token
  schemas:
...
    Episode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        show:
          $ref: '#/components/schemas/ShowSummary'
        title:
          type: string
        slug:
          type: string
        published_at:
          type: string
          format: date
        original_url:
          type: string
        ai_summary:
          type: string
        ai_sponsorship:
          type: string
        host_notes:
          type: string
        duration_seconds:
          type: integer
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        comment_count:
          type: integer
    ShowSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        cover_image_url:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        status:
          type: string
          enum: [pending, approved, hidden, spam]
        created_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type: string
...
    SearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Episode'
        pagination:
          $ref: '#/components/schemas/Pagination'
        filters:
          type: object
          properties:
            shows:
              type: array
              items:
                $ref: '#/components/schemas/ShowSummary'
            tags:
              type: array
              items:
                $ref: '#/components/schemas/Tag'
...
    AdminCommentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Comment'
              - type: object
                properties:
                  episode:
                    $ref: '#/components/schemas/EpisodeSummary'
                  spam_score:
                    type: number
        pagination:
          $ref: '#/components/schemas/Pagination'
    EpisodeSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        show_slug:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer